+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # SAS Viya Monitoring for Kubernetes
 
-## Unreleased
+## UNRELEASED
 * **Metrics**
   * [FIX] Restore metrics after move to Crunchy Data v.5
+  
+* **Logging**
+  * [FIX] Limited Access User Can NOT Generate CSV via OpenSearch Dashboards Reporting
 
 ## Version 1.2.7 (02DEC2022)
+
 * **Logging**
   * [FIX] - Pin to specific Helm chart version (4.15.1) in logging/bin/deploy_esexporter.sh
   * [UPGRADE] - Fluent Bit has been upgraded from version 1.9.6 to 1.9.9